+++ src/main/java/dev/yatloaf/modkrowd/cubekrowd/subserver/MainSubserver.java
@@ -1,10 +1,12 @@
 package dev.yatloaf.modkrowd.cubekrowd.subserver;
 
+import dev.yatloaf.modkrowd.cubekrowd.common.RankName;
 import dev.yatloaf.modkrowd.cubekrowd.common.cache.TextCache;
 import dev.yatloaf.modkrowd.cubekrowd.common.SelfPlayer;
 import dev.yatloaf.modkrowd.util.text.StyledString;
-import net.minecraft.text.Text;
 import net.minecraft.util.Formatting;
+
+import java.util.EnumSet;
 
 public class MainSubserver extends RealSubserver {
 
@@ -14,11 +16,16 @@ public class MainSubserver extends RealSubserver {
 
     @Override
     public TextCache formatChat(String message) {
+        RankName rankName = SelfPlayer.rankNameSoft();
         StyledString prefix = SelfPlayer.rankNameSoft().appearance();
+
         if (StyledString.EMPTY.equals(prefix)) {
             return TextCache.EMPTY;
         } else {
-            return TextCache.of(prefix.toText().append(Text.literal(" " + message).formatted(Formatting.WHITE)));
+            EnumSet<Formatting> permittedFormattings = rankName.rank().letters().permittedFormattings();
+            StyledString formatted = StyledString.fromFormattedString(message, '&', permittedFormattings);
+
+            return TextCache.of(StyledString.concat(prefix, StyledString.SPACE, formatted));
         }
     }
 }